TS-4001: Separate Http2Stream from Http2ConnectionState.h
